sokoyetu/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
PyJWT==2.12.1
cryptography==41.0.7
python-multipart==0.0.6
haversine==2.8.1
qrcode[pil]==7.4.2
pillow==10.1.0
aiofiles==23.2.1
jinja2==3.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.1
africastalking==1.2.5
python-dotenv==1.0.0
bcrypt==4.1.1
pytest==7.4.3
pytest-asyncio==0.21.1

# Database Drivers (optional - choose based on DATABASE_TYPE)
pymysql==1.1.0
psycopg2-binary==2.9.9
